Changed Mode[4] pin from being pulled down to being pulled up on all cores. This is to use the undocumented 'reserved' mode functionality of '110'. Under this mode, all cores (cores 0 & 1 in L2s) will enable boot from Link B AND internal links. Using this, we can boot all cores over links using just the boot ROM.
The only issue this brings up is that the flash connected L2 now has the choice of link boot of BOTH cores booking via SPI. Since only one core (0) is connected to a flash chip, this means that core 1 may be unusable if using boot from Flash. Likely an OTP program will be necessary to disable it from trying. Sub-optimal, but the ability to boot the entire network over links is more important.